refactor: update code quality and test cases
